requirements.txt
Pillow
hachoir
fpdf
image
ttkthemes